# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App",
        "version": "0.1"
    },
    "paths": {
...
        "/api/cms/users": {
            "get": {
                "tags": [
                    "CMS Users"
                ],
                "summary": "Lấy ra danh sách người dùng",
                "operationId": "f1d870a7b8ede0da18c47221287cfabb",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per Page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "searchFields[]",
                        "in": "query",
                        "description": "List of fields to search. Example: ['name_booking']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "name_booking"
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Content search. Example: 'Thiên'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Thiên"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter criteria in JSON format. Example: {\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "{\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}"
                        }
                    },
                    {
                        "name": "sort[]",
                        "in": "query",
                        "description": "Sort criteria in array format. Use '-' for descending order and '+' for ascending order. Example: ['-created_at']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "-created_at"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/users/{id}": {
            "get": {
                "tags": [
                    "CMS Users"
                ],
                "summary": "Lấy thông tin chi tiết người dùng",
                "operationId": "1ce0d13094e0cdc7cfaa8249ac7fbc81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/users/update/{id}": {
            "post": {
                "tags": [
                    "CMS Users"
                ],
                "summary": "Cập nhật người dùng",
                "operationId": "5eb7c0ffb6d799830ed9a78baa6c3938",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name",
                                    "phone_number"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Bánh ngọt"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0387768880"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "422": {
                        "description": "Lỗi validate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}